+++ example/spec/parallel/testAPI/bulkDeleteFailureSpec.js
@@ -43,11 +43,11 @@ describe('POST /granules/bulkDelete with a failed bulk delete operation', () => 
     postBulkDeleteBody = JSON.parse(postBulkDeleteResponse.body);
 
     // Query the AsyncOperation API to get the task ARN
-    const getAsyncOperationResponse = await apiTestUtils.getAsyncOperation({
+    const asyncOperation = await apiTestUtils.getAsyncOperation({
       prefix: config.stackName,
       id: postBulkDeleteBody.id,
     });
-    ({ taskArn } = JSON.parse(getAsyncOperationResponse.body));
+    ({ taskArn } = asyncOperation);
 
     beforeAllSucceeded = true;
   });
@@ -71,16 +71,12 @@ describe('POST /granules/bulkDelete with a failed bulk delete operation', () => 
   it('creates an AsyncOperation', async () => {
     expect(beforeAllSucceeded).toBeTrue();
 
-    const getAsyncOperationResponse = await apiTestUtils.getAsyncOperation({
+    const asyncOperation = await apiTestUtils.getAsyncOperation({
       prefix: config.stackName,
       id: postBulkDeleteBody.id,
     });
 
-    expect(getAsyncOperationResponse.statusCode).toEqual(200);
-
-    const getAsyncOperationBody = JSON.parse(getAsyncOperationResponse.body);
-
-    expect(getAsyncOperationBody.id).toEqual(postBulkDeleteBody.id);
+    expect(asyncOperation.id).toEqual(postBulkDeleteBody.id);
   });
 
   it('runs an ECS task', async () => {
@@ -106,21 +102,18 @@ describe('POST /granules/bulkDelete with a failed bulk delete operation', () => 
       }
     ).promise();
 
-    const getAsyncOperationResponse = await apiTestUtils.getAsyncOperation({
+    const asyncOperation = await apiTestUtils.getAsyncOperation({
       prefix: config.stackName,
       id: postBulkDeleteBody.id,
     });
 
-    const getAsyncOperationBody = JSON.parse(getAsyncOperationResponse.body);
-
-    expect(getAsyncOperationResponse.statusCode).toEqual(200);
-    expect(getAsyncOperationBody.status).toEqual('TASK_FAILED');
+    expect(asyncOperation.status).toEqual('TASK_FAILED');
 
     let output;
     try {
-      output = JSON.parse(getAsyncOperationBody.output);
+      output = JSON.parse(asyncOperation.output);
     } catch (error) {
-      throw new SyntaxError(`getAsyncOperationBody.output is not valid JSON: ${getAsyncOperationBody.output}`);
+      throw new SyntaxError(`getAsyncOperationBody.output is not valid JSON: ${asyncOperation.output}`);
     }
 
     expect(output.name).toBe('AggregateError');