+++ example/spec/parallel/testAPI/bulkDeleteSuccessSpec.js
@@ -165,13 +165,13 @@ describe('POST /granules/bulkDelete', () => {
         postBulkDeleteBody = JSON.parse(postBulkDeleteResponse.body);
 
         // Query the AsyncOperation API to get the task ARN
-        const getAsyncOperationResponse = await apiTestUtils.getAsyncOperation(
+        const asyncOperation = await apiTestUtils.getAsyncOperation(
           {
             prefix,
             id: postBulkDeleteBody.id,
           }
         );
-        ({ taskArn } = JSON.parse(getAsyncOperationResponse.body));
+        ({ taskArn } = asyncOperation);
         beforeAllSucceeded = true;
       } catch (error) {
         console.log(error);
@@ -229,16 +229,12 @@ describe('POST /granules/bulkDelete', () => {
     it('creates an AsyncOperation', async () => {
       expect(beforeAllSucceeded).toBeTrue();
 
-      const getAsyncOperationResponse = await apiTestUtils.getAsyncOperation({
+      const asyncOperation = await apiTestUtils.getAsyncOperation({
         prefix,
         id: postBulkDeleteBody.id,
       });
 
-      expect(getAsyncOperationResponse.statusCode).toEqual(200);
-
-      const getAsyncOperationBody = JSON.parse(getAsyncOperationResponse.body);
-
-      expect(getAsyncOperationBody.id).toEqual(postBulkDeleteBody.id);
+      expect(asyncOperation.id).toEqual(postBulkDeleteBody.id);
     });
 
     it('runs an ECS task', async () => {
@@ -264,21 +260,17 @@ describe('POST /granules/bulkDelete', () => {
         }
       ).promise();
 
-      const getAsyncOperationResponse = await apiTestUtils.getAsyncOperation({
+      const asyncOperation = await apiTestUtils.getAsyncOperation({
         prefix,
         id: postBulkDeleteBody.id,
       });
-
-      const getAsyncOperationBody = JSON.parse(getAsyncOperationResponse.body);
-
-      expect(getAsyncOperationResponse.statusCode).toEqual(200);
-      expect(getAsyncOperationBody.status).toEqual('SUCCEEDED');
+      expect(asyncOperation.status).toEqual('SUCCEEDED');
 
       let output;
       try {
-        output = JSON.parse(getAsyncOperationBody.output);
+        output = JSON.parse(asyncOperation.output);
       } catch (error) {
-        throw new SyntaxError(`getAsyncOperationBody.output is not valid JSON: ${getAsyncOperationBody.output}`);
+        throw new SyntaxError(`asyncOperation.output is not valid JSON: ${asyncOperation.output}`);
       }
 
       expect(output).toEqual({ deletedGranules: [granuleId] });